Set "spanner-share-context" event property with \=.
For events in different voices to be associated together as part of the same
spanner, there must be a common parent context in which data (e.g. Spanner
object) can be shared. Setting the spanner-share-context achieves this.

If only one argument is given to \=, spanner-share-context defaults to Voice
(i.e. created spanners are only visible to the same Voice).