Create multiple engraver instances to handle multiple spanners.
The special macros in spanner-engraver.hh (e.g. ADD_FILTERED_LISTENER) will
first route the event/acknowledgement through the Spanner_engraver class.
Each engraver instance is responsible for a particular spanner-share-context
and spanner-id. The event/acknowledgement is passed to the actual callback
in the Engraver only if it matches the spanner-share-context and spanner-id
for an engraver instance.

To allow this kind of filtering on general acknowledgements, the spanner-id
and spanner-share-context properties may be set on any grob-interface.

Various functions for handling (e.g. adding, ending) cross-voice spanners
via the sharedSpanners context property are also available to classes derived
from Spanner_engraver.